pyproject.toml
[project]
name = "wayonagio-email-agent"
version = "0.1.0"
description = "Email response agent for a Cusco travel agency"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "fastapi>=0.135.1",
    "google-api-python-client>=2.192.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "google-cloud-storage>=3.10.1",
    "litellm>=1.83.0",
    "numpy>=2.4.4",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]

[project.scripts]
wayonagio = "wayonagio_email_agent.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = [
    # Treat ResourceWarning as a test failure: the SQLite connection leak in
    # state.py and kb/store.py both surfaced this way and silently passed CI
    # for weeks before we caught them. Locking it down here means any new
    # leak (file handle, socket, sqlite connection) breaks the suite the
    # moment it's introduced.
    "error::ResourceWarning",
    # Starlette's BaseHTTPMiddleware internals still reference the pre-RFC9110
    # name for HTTP 413. Our own code uses HTTP_413_CONTENT_TOO_LARGE; this
    # filter keeps the test output focused on issues we can actually fix.
    "ignore:.*HTTP_413_REQUEST_ENTITY_TOO_LARGE.*:DeprecationWarning",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
